requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.3
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.19.1
pyarrow==6.0.0
pycountry==20.7.3
pycparser==2.20
pydeck==0.7.1
Pygments==2.10.0
pymongo==3.12.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.26.0
scikit-learn==0.22.1
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.1.0
streamlit-lottie==0.0.2
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
